move LB policy registry construction into bootstrap builder
