+++ src/core/xds/grpc/xds_lb_policy_registry.h
@@ -17,8 +17,6 @@
 #ifndef GRPC_SRC_CORE_XDS_GRPC_XDS_LB_POLICY_REGISTRY_H
 #define GRPC_SRC_CORE_XDS_GRPC_XDS_LB_POLICY_REGISTRY_H
 
-#include <grpc/support/port_platform.h>
-
 #include <map>
 #include <memory>
 
@@ -45,7 +43,12 @@ class XdsLbPolicyRegistry final {
     virtual absl::string_view type() = 0;
   };
 
-  XdsLbPolicyRegistry();
+  XdsLbPolicyRegistry() = default;
+
+  template <typename T>
+  void RegisterFactory(std::unique_ptr<T> factory) {
+    policy_config_factories_.emplace(T::Type(), std::move(factory));
+  }
 
   // Converts an xDS cluster load balancing policy message to gRPC's JSON
   // format. An error is returned if none of the lb policies in the list are
